examples/SO2SAT_RAW_format/requirements.txt
tensorflow==2.7.0
tensorflow-datasets==4.5.2
kafka-python==2.0.2